backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.5
django-cors-headers==4.6.0
django-environ==0.12.0
djangorestframework==3.15.2
idna==3.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
requests==2.32.3
ruff==0.9.5
shortuuid==1.0.13
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
